+++ MolAligner/__init__.py
@@ -3,6 +3,7 @@
 from .version import __version__
 from .molecule import Molecule
 from .aligner import Aligner
+from .alignerx import AlignerX
 from . import rotation_matrix
 from . import units
 from . import utility
