+++ tests/test_replicate.py
@@ -1,7 +1,8 @@
-from MolAligner import Aligner
+from MolAligner import AlignerX
 
-mol = Aligner("MFI.pdb")
+mol = AlignerX("MFI.pdb")
 a, b, c = 20.0900, 19.7380, 13.1420
 box = [a, b, c]
-mol.pbc_replicate(box, multiple=(7, -7, 3))
+mol.set_box(box)
+mol.pbc_replicate(multiple=(7, -7, 3))
 mol.write("MFI-replicate.gro")
